requirements.txt
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
distro==1.9.0
flatbuffers==25.12.19
google==3.0.0
google-auth==2.49.2
google-genai==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio==2.37.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-loader==0.5
llvmlite==0.47.0
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
onnxruntime==1.24.4
onnxruntime-gpu==1.24.4
packaging==26.1
pillow==12.2.0
pip==25.3
platformdirs==4.9.6
pooch==1.9.0
protobuf==7.34.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-core==2.46.3
pymatting==1.1.15
python-dotenv==1.2.2
referencing==0.37.0
rembg==2.0.75
requests==2.33.1
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
sniffio==1.3.1
soupsieve==2.8.3
sympy==1.14.0
tenacity==9.1.4
tifffile==2026.4.11
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uv==0.11.7
websockets==16.0
